Dolor, Refactor
todo separado en escenas me cago en dios
