+++ src/javafxmlapplication/pistas/FXMLPistas.fxml
@@ -1,0 +1,42 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.DatePicker?>
+<?import javafx.scene.control.ScrollPane?>
+<?import javafx.scene.layout.BorderPane?>
+<?import javafx.scene.layout.FlowPane?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.Pane?>
+
+
+<BorderPane xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="javafxmlapplication.pistas.FXMLPistasController">
+   <top>
+      <HBox alignment="CENTER" prefHeight="100.0" prefWidth="200.0" BorderPane.alignment="CENTER">
+         <children>
+            <Button fx:id="antButton" mnemonicParsing="false" onAction="#onAntButton" text="Anterior" HBox.hgrow="NEVER" />
+            <Pane prefHeight="200.0" HBox.hgrow="ALWAYS" />
+            <DatePicker fx:id="dpBookingDay" />
+            <Pane prefHeight="200.0" HBox.hgrow="ALWAYS" />
+            <Button fx:id="posButton" mnemonicParsing="false" onAction="#onPosButton" text="Siguiente" HBox.hgrow="NEVER" />
+         </children>
+         <BorderPane.margin>
+            <Insets />
+         </BorderPane.margin>
+      </HBox>
+   </top>
+   <center>
+      <ScrollPane fx:id="scrollPane" BorderPane.alignment="CENTER">
+         <content>
+            <HBox alignment="TOP_CENTER">
+               <children>
+                  <FlowPane fx:id="flowPane" columnHalignment="CENTER" prefWidth="1915.0" />
+               </children>
+            </HBox>
+         </content>
+         <BorderPane.margin>
+            <Insets />
+         </BorderPane.margin>
+      </ScrollPane>
+   </center>
+</BorderPane>
