+++ src/javafxmlapplication/reservas/FXMLReservas.fxml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.ScrollPane?>
+<?import javafx.scene.layout.BorderPane?>
+<?import javafx.scene.layout.VBox?>
+
+
+<BorderPane xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="javafxmlapplication.reservas.FXMLReservasController">
+   <top>
+      <Label fx:id="nombreUsuarioReservas" alignment="CENTER" contentDisplay="CENTER" prefHeight="63.0" prefWidth="413.0" text="Label" BorderPane.alignment="CENTER" />
+   </top>
+   <center>
+      <ScrollPane fx:id="misReservasScrollPane" BorderPane.alignment="CENTER">
+         <content>
+            <VBox fx:id="misReservasContainer" alignment="TOP_CENTER" prefHeight="100.0" prefWidth="1000.0" />
+         </content>
+      </ScrollPane>
+   </center>
+</BorderPane>
